Continued work on file sys calls

+++ kern/include/fd.h
@@ -31,6 +31,7 @@
 #define _FD_H_
 
 #include <limits.h>
+#include <types.h>
 
 /*
  * File handle.
@@ -40,7 +41,7 @@
 struct fd {
 	//char name[FILENAME_MAX];	/* Maximum file name length */
 	int flags;					/* Flags passed in by open() */
-	int offset;				/* Current seek position */
+	off_t offset;				/* Current seek position */
 	int ref_count;
 	struct lock* lock;			/* For mutual exclusion. */
 	struct vnode* vn;			/* Interfacing with vnode layer. */
